Added files for HW-Oct-06
P238/13. For testing the preceding logic we need numeric value.
14. The fundamental rule of the testing method is every method must be tested within a method before testing the parent method.

15. Following the style guidelines given in this book, when should an instance
variable be marked private?
Ans: It's a good practice to use every instance variable as private and use other public methods to initialize them.

16. Following the style guidelines given in this book, when should a method be
marked private?
Ans: When we do not want to be it visible outside the class. Also, It is to be used by other methods internally.

17. List all the accessor methods in the class DateFifthTry
Ans: getDay, getYear, getMonth, monthString

18. List all the mutator methods in the class DateFifthTry
Ans: setDate, setMonth, setDay, setYear,